Moved `LiveIterableWrapper` to utils and added default constructors.
